Move the unsubscribe url helper
